### 2020-03-26 #23 @mluerig
Phenopype: a phenotyping pipeline for Python
Submitting Author: Name (@mluerig)
Package Name: Phenopype
One-Line Description of Package: a phenotyping pipeline for Python
Repository Link (if existing):  https://github.com/mluerig/phenopype

---

Phenopype is a high throughput phenotyping pipeline for Python that aims at supporting biologists in their efforts to extract high dimensional phenotypic data from digital images. Phenopype provides high level functions for image processing that can be stacked and executed sequentially to efficiently process single images or large data sets in a semi or fully automated fashion. Users can assemble their own function-stacks that can be customized and stored along with raw data for full reproducibility (check the high throughput workflow). Phenopype can be run from Python or from a Python Integrated Development Environment (IDE), like Spyder. Some Python knowledge is necessary, but most of the heavy lifting is done in the background. Phenopype can be installed from the Python Package Index (PYPI) using pip install phenopype.

- Please indicate which [category or categories](https://pyopensci.github.io/dev_guide/peer_review/peer_review_proc.html#aims-and-scope) this package falls under:

Phenopype is designed to extract phenotypic data (https://en.wikipedia.org/wiki/Phenotype) of plants, animals, and other organisms from images and videos.

Phenopype is intended for ecologists and evolutionary biologists that work with phenotypic data. Phenotypic data are an essential component of ecological and evolutionary research (https://www.nature.com/articles/nrg2897)

Only low level computer vision packages like OpenCV or scikit-image are out there that require a lot of configuring and a relatively deep understanding of computer vision and Python in general. Phenopype offers high level functions so that users can focus on the relevant analytic parts of image analysis.

Documentation (https://mluerig.github.io/phenopype/) is semi-complete. I am working on finishing up all docstrings and making the package PEP8 conform.

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*

EDIT: fixed typos
