### 2020-03-27 aa18a2a
docstr main/utils
