+++ docs/conf.py
@@ -40,7 +40,8 @@ extensions = [
     'sphinx.ext.coverage',
     'sphinx.ext.napoleon',
     'sphinx.ext.githubpages',
-    "nbsphinx"
+    "nbsphinx",
+    'sphinx.ext.autosectionlabel'
 ]
 
 autodoc_member_order = 'bysource'
@@ -96,7 +97,7 @@ html_theme_options = {
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = []
+html_static_path = ["images","_images"]
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
