+++ docs/_sources/example_2_landmarks_stickleback.ipynb.txt
@@ -15,13 +15,21 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "<center>\n",
-    "<div style=\"width:600px; text-align: left\" >\n",
+    "<div class=\"row; text-align: left\">\n",
     "    \n",
-    "![Phenopype workflow example](images/stickle1.JPG)\n",
+    "<div class=\"col-md-6\">\n",
     "    \n",
+    "![Before](_assets/stickle1.jpg)\n",
+    "    \n",
+    "**Input** - Stained threespine stickleback, photographed in a glycerol bath from a camera stand\n",
     "</div>\n",
-    "</center>"
+    "<div class=\"col-md-6\">\n",
+    "\n",
+    "![After](_assets/stickle1_lm.jpg)\n",
+    "    \n",
+    "**Results** - 22 landmarks are placed using the `landmark` tool from the `phenopype.measurements` module\n",
+    "</div>\n",
+    "</div>\n"
    ]
   },
   {
@@ -33,7 +41,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 1,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -46,7 +54,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 2,
    "metadata": {},
    "outputs": [
     {
@@ -76,6 +84,94 @@
   },
   {
    "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/html": [
+       "<div>\n",
+       "<style scoped>\n",
+       "    .dataframe tbody tr th:only-of-type {\n",
+       "        vertical-align: middle;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe tbody tr th {\n",
+       "        vertical-align: top;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe thead th {\n",
+       "        text-align: right;\n",
+       "    }\n",
+       "</style>\n",
+       "<table border=\"1\" class=\"dataframe\">\n",
+       "  <thead>\n",
+       "    <tr style=\"text-align: right;\">\n",
+       "      <th></th>\n",
+       "      <th>filename</th>\n",
+       "      <th>width</th>\n",
+       "      <th>height</th>\n",
+       "      <th>size_ratio_original</th>\n",
+       "      <th>landmark</th>\n",
+       "      <th>x</th>\n",
+       "      <th>y</th>\n",
+       "    </tr>\n",
+       "  </thead>\n",
+       "  <tbody>\n",
+       "    <tr>\n",
+       "      <th>0</th>\n",
+       "      <td>stickle1.jpg</td>\n",
+       "      <td>2400</td>\n",
+       "      <td>1600</td>\n",
+       "      <td>1</td>\n",
+       "      <td>1</td>\n",
+       "      <td>1034</td>\n",
+       "      <td>451</td>\n",
+       "    </tr>\n",
+       "    <tr>\n",
+       "      <th>1</th>\n",
+       "      <td>stickle1.jpg</td>\n",
+       "      <td>2400</td>\n",
+       "      <td>1600</td>\n",
+       "      <td>1</td>\n",
+       "      <td>2</td>\n",
+       "      <td>1202</td>\n",
+       "      <td>326</td>\n",
+       "    </tr>\n",
+       "    <tr>\n",
+       "      <th>2</th>\n",
+       "      <td>stickle1.jpg</td>\n",
+       "      <td>2400</td>\n",
+       "      <td>1600</td>\n",
+       "      <td>1</td>\n",
[…]
