+++ docs/conf.py
@@ -97,7 +97,7 @@ html_theme_options = {
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = []
+html_static_path = ["../tutorials/_assets", "../_assets"]
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
