+++ docs/_sources/index.rst.txt
@@ -5,7 +5,7 @@ Phenopype: a phenotyping pipeline for Python
 
 	<hr>
 
-	<img src="_images/logo.png">
+	<img src="../_assets/logo.png">
 
 	<div align="justify">
 	<br>
