+++ docs/_sources/index.rst.txt
@@ -5,13 +5,17 @@ Phenopype: a phenotyping pipeline for Python
 
 	<hr>
 
-	<img src="../_assets/logo.png">
+.. image:: logo.png
+
+.. raw:: html
 
 	<div align="justify">
 	<br>
-	<strong>Phenopype is a high throughput phenotyping pipeline for Python that aims at supporting biologists in their efforts to extract high dimensional phenotypic data from digital images</strong>. Phenopype provides high level functions for image processing that can be stacked and executed sequentially to efficiently process single images or large data sets in a semi or fully automated fashion. Users can assemble their own function-stacks that can be customized and stored along with raw data for full reproducibility (check the high throughput workflow). Phenopype can be run from Python or from a Python Integrated Development Environment (IDE), like <a href="https://www.spyder-ide.org/">Spyder</a>. Some Python knowledge is necessary, but most of the heavy lifting is done in the background. Phenopype can be installed from the Python Package Index (PYPI) using <code>pip install phenopype</code>. To get started, <a href="https://github.com/mluerig/phenopype/archive/master.zip">clone the Github-repo</a> and <a href="tutorial_0.html">run the tutorials</a>. 
+	<strong>Phenopype is a high throughput phenotyping pipeline for Python that aims at supporting biologists in their efforts to extract high dimensional phenotypic data from digital images</strong>. Phenopype provides high level functions for image processing that can be stacked and executed sequentially to efficiently process single images or large data sets in a semi or fully automated fashion. Users can assemble their own function-stacks that can be customized and stored along with raw data for full reproducibility (check the high throughput workflow). Phenopype can be run from Python or from a Python Integrated Development Environment (IDE), like <a href="https://www.spyder-ide.org/">Spyder</a>. Some Python knowledge is necessary, but most of the heavy lifting is done in the background. Phenopype can be installed from the Python Package Index (PYPI) using <code>pip install phenopype</code>. 
+	<br>
+	<br>
+	To get started, <a href="https://github.com/mluerig/phenopype/archive/master.zip">clone the Github-repo</a> and <a href="tutorial_0.html">run the tutorials</a>. 
 	</div>
-
 	<hr>
 
 	<div class="container-fluid">
