+++ docs/_sources/example_2_landmarks_stickleback.ipynb.txt
@@ -19,13 +19,13 @@
     "    \n",
     "<div class=\"col-md-6\">\n",
     "    \n",
-    "![Before](../_assets/stickle1.jpg)\n",
+    "![Before](_assets/stickle1.jpg)\n",
     "    \n",
     "**Input** - Stained threespine stickleback, photographed in a glycerol bath from a camera stand\n",
     "</div>\n",
     "<div class=\"col-md-6\">\n",
     "\n",
-    "![After](../_assets/stickle1_lm.jpg)\n",
+    "![After](_assets/stickle1_lm.jpg)\n",
     "    \n",
     "**Results** - 22 landmarks are placed using the `landmark` tool from the `phenopype.measurements` module\n",
     "</div>\n",
@@ -201,7 +201,7 @@
     "<center>\n",
     "<div style=\"width:500px; text-align: left\">\n",
     "    \n",
-    "![Adding a scale](../_assets/set_scale.gif)\n",
+    "![Adding a scale](_assets/set_scale.gif)\n",
     "    \n",
     "</div>\n",
     "</center>"
@@ -469,8 +469,9 @@
     "<center>\n",
     "<div style=\"text-align: left\">\n",
     "    \n",
-    "![Landmarks in high throughput](../_assets/landmarks_high_throughput.gif)\n",
-    "\n",
+    "![Landmarks in high throughput](_assets/landmarks_high_throughput.gif)\n",
+    "    \n",
+    "**High throughput method** - the `pype` method opens the image and a text editor with the pype configuration file. Any change to the configuration file, in this case the point size for landmarks, will be immediately applied to the image.  \n",
     "</div>\n",
     "</center>"
    ]
