+++ docs/_sources/resources.rst.txt
@@ -20,10 +20,9 @@ Tutorials
 
 Helpful questions on StackOverflow
 
-- https://stackoverflow.com/questions/20981789/difference-between-methods-and-functions-in-python-compared-to-c
-- https://stackoverflow.com/questions/155609/whats-the-difference-between-a-method-and-a-function?noredirect=1&lq=1
+- https://stackoverflow.com/questions/155609/whats-the-difference-between-a-method-and-a-function
 - https://stackoverflow.com/questions/509211/understanding-slice-notation
-- https://stackoverflow.com/questions/1977362/how-to-create-module-wide-variables-in-python?noredirect=1&lq=1
+- https://stackoverflow.com/questions/1977362/how-to-create-module-wide-variables-in-python
 - https://stackoverflow.com/questions/448271/what-is-init-py-for
 - https://stackoverflow.com/questions/16476924/how-to-iterate-over-rows-in-a-dataframe-in-pandas
 
