+++ docs/_sources/tutorial_2_phenopype_workflow.ipynb.txt
@@ -55,7 +55,7 @@
    "source": [
     "The low throughput workflow starts with the path to an image that is stored on the hard drive. `load_image` imports the file as a three-channel [1] numpy array (*ndarray*), together with image meta data (file name, exposure, dimensions, etc.) as a pandas *DataFrame*. The array gets passed on to the `threshold` function, which will return a binary array of the same dimensions. This array needs to be passed on to the `find_contours` function, which will return a dictionary with the detected contours. This dictionary, together with the original array, can then be passed to the `colour_intensity` function. This function will collect the average color value from within the perimeter coordinates for each contour and return a pandas dataframe containing those values. Finally, the dataframe can be exported as a csv file with `save_colour`. By passing on the initially created meta-data, the function will automatically expand the provided columns of meta-info into the exported csv.  \n",
     "\n",
-    "[1] to learn more about the basic of Computer Vision check the [resources section](resources.html) of the phenopype documentation. "
+    "[1] to learn more about the basics of Computer Vision check the [resources section](resources.html) of the phenopype documentation. "
    ]
   },
   {
@@ -74,7 +74,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 1,
    "metadata": {},
    "outputs": [
     {
@@ -83,7 +83,7 @@
      "text": [
       "- create mask\n",
       "- applying mask: mask1\n",
-      "- contours saved under ../_temp/output\\contours.csv (overwritten).\n"
+      "- contours saved under ../_temp/output\\contours.csv.\n"
      ]
     }
    ],
@@ -125,7 +125,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": null,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -155,7 +155,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 1,
    "metadata": {},
    "outputs": [
     {
@@ -164,7 +164,7 @@
      "text": [
       "- create mask\n",
       "- applying mask: mask1\n",
-      "- contours saved under images\\contours.csv.\n"
+      "- contours saved under ../_temp/output\\contours.csv.\n"
      ]
     }
    ],
@@ -199,7 +199,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": null,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -215,17 +215,9 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 10,
+   "execution_count": null,
    "metadata": {},
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "['__class__', '__delattr__', '__dict__', '__dir__', '__doc__', '__eq__', '__format__', '__ge__', '__getattribute__', '__gt__', '__hash__', '__init__', '__init_subclass__', '__le__', '__lt__', '__module__', '__ne__', '__new__', '__reduce__', '__reduce_ex__', '__repr__', '__setattr__', '__sizeof__', '__str__', '__subclasshook__', '__weakref__', 'canvas', 'df_contours', 'df_image_data', 'df_image_data_copy', 'df_masks', 'dirpath', 'image', 'image_bin', 'image_copy', 'image_data', 'image_gray', 'image_mod', 'load', 'reset', 'save', 'save_suffix', 'show']\n"
-     ]
-    }
-   ],
+   "outputs": [],
    "source": [
     "print(dir(container))"
    ]
@@ -241,7 +233,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
[…]
