### 2020-04-03 c1056d4
creat scale add mask
