### 2020-04-05 4bb9cfe
fix ex1 ass functions
