+++ docs/_sources/example_1_detect_objects_isopods.ipynb.txt
@@ -6,7 +6,991 @@
    "source": [
     "# Example 1: Measuring shape, size and colour of isopods  \n",
     "\n",
-    "[ADD TEXT]"
+    "Little nugget of information: this was the original computer vision problem that phenopype was intended to solve (it's predecessor [\"iso_cv\"](https://github.com/mluerig/iso_cv) was not much more than a script-collection).\n",
+    "\n",
+    "In this example I am demonstrating all three phenopype workflows ([prototyping](#prototyping), [low thoughput](#low-throughput) and [high throughput](#high-throughput)). For a project, you would probably only use either the high or low throughput approach, but the protoypting approach for this example nicely decomposes the all the required steps for a classic computer vision workflow. "
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "<div class=\"row; text-align: left\">\n",
+    "    \n",
+    "<div class=\"col-md-6\">\n",
+    "    \n",
+    "![Before](_assets/ex1_before.jpg)\n",
+    "    \n",
+    "**Input** - Freshwater isopod, alive, photographed on a white resin-tray from a camera stand. \n",
+    "</div>\n",
+    "<div class=\"col-md-6\">\n",
+    "\n",
+    "![After](_assets/ex1_after.jpg)\n",
+    "    \n",
+    "**Results** - Isopod shape, size and colour are extracted (and size referenced using the reference card) \n",
+    "</div>\n",
+    "</div>"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Prototyping"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Loading the image\n",
+    "\n",
+    "First we import the image from the a filepath using `load_image`. With the flag `df=True` we can extract some image-meta information that we want to be represented in the results files later (i.e. image name and its dimensions). After loading it, we can have a quick look at it with the `show_image` function - you can close it again with Enter or Esc. "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import phenopype as pp"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "filepath = r\"images/isopods.jpg\"\n",
+    "image, img_data = pp.load_image(filepath, df=True)\n",
+    "pp.show_image(image)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/html": [
+       "<div>\n",
+       "<style scoped>\n",
+       "    .dataframe tbody tr th:only-of-type {\n",
+       "        vertical-align: middle;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe tbody tr th {\n",
+       "        vertical-align: top;\n",
+       "    }\n",
+       "\n",
+       "    .dataframe thead th {\n",
+       "        text-align: right;\n",
+       "    }\n",
+       "</style>\n",
+       "<table border=\"1\" class=\"dataframe\">\n",
+       "  <thead>\n",
+       "    <tr style=\"text-align: right;\">\n",
+       "      <th></th>\n",
+       "      <th>filename</th>\n",
+       "      <th>width</th>\n",
+       "      <th>height</th>\n",
+       "      <th>size_ratio_original</th>\n",
+       "    </tr>\n",
+       "  </thead>\n",
[…]
