+++ docs/_sources/tutorial_0.rst.txt
@@ -1,5 +1,5 @@
-How to run the tutorials
-========================
+How to run the tutorials and examples using Jupyter Notebook
+============================================================
 
 The tutorials are written as `Jupyter notebooks`, interactive web-apps to run, document, and visualize code: `https://jupyter.org/ <https://jupyter.org/>`_. If you installed Python using Anaconda, it is possible that :code:`jupyter` is already installed - check py typing :code:`jupyter --version` into the console. Otherwise you can install it with pip
 
@@ -15,3 +15,29 @@ After installing phenopype, download and unpack the phenopype repository from `g
 
 Alternatively, you can copy the code from the jupyter code cells over to Spyder and execute them there. In this case you need to adjust the file paths given in the code cells accordingly. 
 
+
+Tutorials
+---------
+
+.. toctree::
+	:maxdepth: 1
+
+	tutorial_1_python_intro
+	tutorial_2_phenopype_workflow
+	tutorial_3_managing_projects_1
+	tutorial_4_managing_projects_2
+	tutorial_5_gui_interactions
+	tutorial_6_video_analysis
+
+Examples
+-------- 
+
+.. toctree::
+	:maxdepth: 1
+
+	example_1_detect_objects_isopods
+	example_2_landmarks_stickleback
+	example_3_phytoplankton
+	example_4_video_analysis_stickleback
+	example_5_shape_stickleback
+	example_6_counting_snails
