+++ docs/_sources/example_3_phytoplankton.ipynb.txt
@@ -6,7 +6,9 @@
    "source": [
     "# Example 3: Fluorescence intensity and shape of phytoplankton cells\n",
     "\n",
-    "[ADD TEXT]"
+    "[ADD TEXT]\n",
+    "\n",
+    "Image kindly provided by Irene Gallego and Anita Narwani."
    ]
   },
   {
