### 2020-04-07 #23 @lwasser
hey @mluerig can you provide a bit more explanation about the exact functionality that phenotype provides? i was a bit unclear after looking at the docs! many thanks!
