### 2020-04-07 #23 @mluerig
I guess the documentation is still a bit confusing. In short phenopype aims at providing a comprehensive and easy-to-use high throughput image analysis workflow using classic computer vision (no machine learning - yet). It aims at ecologists and evolutionary biologists that want to quickly analyze images of organisms and extract phenotypic data.

The provided functions span image data set management (through projects), preprocessing (e.g. setting a common size and color reference and correct images), segmentation (e.g. thresholding and watershed), measurement (e.g. pixel-intensity or landmarks), as well as visualizing and exporting the produced results. The provided functions are designed to be intuitive and intend to minimize user interaction and manual work as much as possible. So, everything is streamlined towards getting solid results fast, even if you don't have a strong programming background.

The other idea is that all settings, intermediate image analysis steps, (e.g. the contour of detected objects) and raw data are available after the analysis so that scientists can generate "cookie-cutter" methods that can be reused, and shared. This will also allow reviewers to reproduce the obtained results with a single line, so nobody has to dig through complex scripts - this makes the collected data more reproducible, which is becoming more and more important.

Does that makes it a bit more clear? Please let me know where the documentation is unclear. Also, there is a written manuscript that provides more comprehensive information - let me know if you would like to see it.
