### 2020-04-07 #23 @lwasser
thank you @mluerig !! i will get back to you. The challenge I am facing right now is just ensuring that we at pyopensci have reviewers who can effectively review packages that are on the analytics side of things! if you were to submit this, can you think of 2 people who would have the skills required to review? i also can ask around on twitter but to be transparent, we are trying to decide what packages we can support review of now and so that is something i consider when a package comes in to us. thank you so much for the speedy response! and again my apologies for such a slow response time.
