### 2020-04-07 #23 @lwasser
thank you @mluerig give me a little bit of time.
What i'm trying to sess out is how analytics focused this package is.
The steps will be
1. you submit a full submission rather than a pre submission as this one is! i will ask you to do this if i can get a bit more info from our team.
2. I will then ask you about suggesting reviewers and we will ping them here.

we have a meeting coming up on thursday if you  have a chance to attend. i plan to bring up this and another presubmission package to ensure it's "in scope". Just so yo uknow my only concern has been analytics focused packages can be difficult to quality check given so few people have the expertise needed. However on the other hand, if we can find people with sufficient expertise, and trust in our reviewers similar to journals I think we should consider a broader range of packages . if you can hang tight until thursday, that would be great. you are also WELCOME to join us at 11am mountain time which i know might be late for you so i understand if that doesn't work!
