### 2020-04-07 #23 @mluerig
okay no problem it can wait. I'm also happy to join the meeting if it helps sorting things out (my email is [here](https://www.eawag.ch/en/aboutus/portrait/organisation/staff/profile/moritz-luerig/show/))
