### 2020-04-09 #23 @mluerig
just let me know how I can join, then I'll try to make it (I'll be in the mountains all day)
