### 2020-04-09 #23 @lwasser
https://pyopensci.discourse.group/t/april-9-community-meeting/169 if you login to our discourse (you can use your github login!) you will have the meeting information. sorry this took me a bit of time to get to -- was struggling with how to avoid being "bombed" online in a meeting!!
