+++ docs/_sources/index.rst.txt
@@ -11,10 +11,14 @@ Phenopype: a phenotyping pipeline for Python
 
 	<div align="justify">
 	<br>
-	<strong>Phenopype is a high throughput phenotyping pipeline for Python that aims at supporting biologists in their efforts to extract high dimensional phenotypic data from digital images</strong>. Phenopype provides high level functions for image processing that can be stacked and executed sequentially to efficiently process single images or large data sets in a semi or fully automated fashion. Users can assemble their own function-stacks that can be customized and stored along with raw data for full reproducibility (check the high throughput workflow). Phenopype can be run from Python or from a Python Integrated Development Environment (IDE), like <a href="https://www.spyder-ide.org/">Spyder</a>. Some Python knowledge is necessary, but most of the heavy lifting is done in the background. Phenopype can be installed from the Python Package Index (PYPI) using <code>pip install phenopype</code>. 
+	<strong>Phenopype is a high throughput phenotyping pipeline for Python that aims at supporting biologists in their efforts to extract high dimensional phenotypic data from digital images</strong>. Phenopype provides high level image-preprocessing, -segmentation, and -analysis functions that can be executed sequentially in a semi or fully automated fashion. Users can assemble their own function-stacks that can be customized and stored along with raw data for full reproducibility. Phenopype can be run from Python or from a Python Integrated Development Environment (IDE), like <i>Spyder</i>. Phenopype is designed to provide robust image analysis workflows that can be implemented with little or no Python experience.<br>
 	<br>
-	<br>
-	To get started, <a href="https://github.com/mluerig/phenopype/archive/master.zip">clone the Github-repo</a> and <a href="tutorial_0.html">run the tutorials</a>. 
+	<strong>Getting started:</strong>
+		<ol>
+			<li><a href="installation.html">Install Phenopype </a> - via the <i>Python Package Index</i> (PYPI): <code>pip install phenopype</code></li> 
+			<li><a href="tutorial_0.html">Run the Tutorials </a> - Tutorial 1 is for Python beginners, otherwise Tutorial 2 is a good starting point </li>
+			<li><a href="#examples">Check the Examples</a> - Example 1 delineates a typical computer vision workflow </li>
+		</ol>
 	</div>
 	<hr>
 
@@ -51,9 +55,12 @@ Phenopype: a phenotyping pipeline for Python
 
 	</div>
 	</div>
-
 	<hr>
-	<h2>Examples</h2>
+
+Examples
+--------
+
+.. raw:: html
 
 	<div class="row">
 	
@@ -61,7 +68,7 @@ Phenopype: a phenotyping pipeline for Python
 			<div class="thumbnail">
 			<a href="example_1_detect_objects_isopods.html">
 			<img src="_images/thumbs/ex1.jpg" alt="example_1_detect_objects_isopods" style="width:100%">
-			<b><p style="text-decoration: underline;">Detect invertebrates</p></b>
+			<b><p style="text-decoration: underline;">Detecting invertebrates</p></b>
 			<b>Organism:</b> freshwater isopods <br>
 			<b>Species:</b> <i>Asellus aquaticus</i> <br>
 			<b>Traits:</b> size, pigmentation, body-shape <br>
@@ -73,7 +80,7 @@ Phenopype: a phenotyping pipeline for Python
 			<div class="thumbnail">
[…]
