+++ tests/resources/project2/data/0__stickle2/attributes.yaml
@@ -1,0 +1,15 @@
+image:
+  filename: stickle2.JPG
+  filepath: tutorials/images\stickle2.JPG
+  filetype: .JPG
+  width: 2400
+  height: 1600
+  size_ratio_original: 1
+project:
+  dirname: 0__stickle2
+  dirpath: tests/resources/project2\data\0__stickle2
+  raw_mode: link
+  raw_path: tutorials/images\stickle2.JPG
+scale:
+  template_path: tests/resources/project2\scale_template.jpg
+  template_px_mm_ratio: 36
