+++ tests/resources/project2/data/0__stickleback_side/attributes.yaml
@@ -1,0 +1,23 @@
+image:
+  filename: stickleback_side.jpg
+  filepath: tutorials/images\stickleback_side.jpg
+  filetype: .jpg
+  width: 3000
+  height: 2000
+  size_ratio_original: 1
+project:
+  dirname: 0__stickleback_side
+  dirpath: tests/resources/project2\data\0__stickleback_side
+  raw_mode: link
+  raw_path: tutorials/images\stickleback_side.jpg
+scale:
+  template_path: tests/resources/project2\scale_template.jpg
+  template_px_mm_ratio: 36
+  current_px_mm_ratio: 36.0
+drawing:
+  tool: line
+  line_width: 2
+  colour: black
+  coords: '[[(1585, 721), (1597, 772)], [(1618, 715), (1610, 769)]]'
+other:
+  ID: '142501'
