+++ tests/resources/project1/attributes.yaml
@@ -1,0 +1,4 @@
+date_created: '20200416_191106'
+date_changed: '20200416_191106'
+root_dir: tests/resources/project1
+data_dir: tests/resources/project1\data
