+++ tests/resources/project1/data/0__stickleback_side/pype_config_v1.yaml
@@ -8,7 +8,7 @@ image:
 pype:
   name: v1
   preset: demo1
-  date_created: '20200416_191106'
+  date_created: '20200417_104807'
 preprocessing:
 - create_mask:
     tool: polygon
