+++ tests/resources/project1/attributes.yaml
@@ -1,4 +1,4 @@
-date_created: '20200416_191106'
-date_changed: '20200416_191106'
+date_created: '20200417_104807'
+date_changed: '20200417_104807'
 root_dir: tests/resources/project1
 data_dir: tests/resources/project1\data
