+++ tests/resources/project2/data/0__stickleback_side/pype_config_v1.yaml
@@ -8,7 +8,7 @@ image:
 pype:
   name: v1
   preset: demo1
-  date_created: '20200415_185858'
+  date_created: '20200417_104808'
 preprocessing:
 - create_mask:
     tool: polygon
@@ -28,15 +28,15 @@ segmentation:
     min_diameter: 0
     min_area: 150
 measurement:
-- polylines
 visualization:
 - select_canvas:
-    canvas: raw
-- draw_contours
-- draw_landmarks
+    canvas: image
+- draw_contours:
+    line_width: 2
+    label_width: 1
+    label_size: 1
+    fill: 0.3
 - draw_masks
-- draw_polylines
-
 export:
 - save_contours:
     overwrite: true
