+++ tests/resources/project2/data/0__stickle2/pype_config_v1.yaml
@@ -8,7 +8,7 @@ image:
 pype:
   name: v1
   preset: demo1
-  date_created: '20200415_185858'
+  date_created: '20200417_104808'
 preprocessing:
 - create_mask:
     tool: polygon
