+++ tests/resources/project2/attributes.yaml
@@ -1,4 +1,4 @@
-date_created: '20200417_104807'
-date_changed: '20200417_104807'
+date_created: '20200415_185858'
+date_changed: '20200415_185858'
 root_dir: tests/resources/project2
 data_dir: tests/resources/project2\data
