+++ tests/resources/project1/data/0__stickle1/attributes.yaml
@@ -1,12 +1,12 @@
 image:
   filename: stickle1.JPG
-  filepath: tutorials/images\stickle1.JPG
+  filepath: tutorials/images/stickle1.JPG
   filetype: .JPG
   width: 2400
   height: 1600
   size_ratio_original: 1
 project:
   dirname: 0__stickle1
-  dirpath: tests/resources/project1\data\0__stickle1
+  dirpath: tests/resources/project1/data/0__stickle1
   raw_mode: link
-  raw_path: tutorials/images\stickle1.JPG
+  raw_path: tutorials/images/stickle1.JPG
