+++ tests/resources/project1/data/0__stickleback_side/pype_config_v1.yaml
@@ -1,6 +1,6 @@
 image:
   filename: stickleback_side.jpg
-  filepath: tutorials/images\stickleback_side.jpg
+  filepath: tutorials/images/stickleback_side.jpg
   filetype: .jpg
   width: 3000
   height: 2000
@@ -8,7 +8,7 @@ image:
 pype:
   name: v1
   preset: demo1
-  date_created: '20200416_191106'
+  date_created: '20200417_122433'
 preprocessing:
 - create_mask:
     tool: polygon
