+++ tests/resources/project1/data/0__stickle3/pype_config_v1.yaml
@@ -1,6 +1,6 @@
 image:
   filename: stickle3.JPG
-  filepath: tutorials/images\stickle3.JPG
+  filepath: tutorials/images/stickle3.JPG
   filetype: .JPG
   width: 2400
   height: 1600
@@ -8,7 +8,7 @@ image:
 pype:
   name: v1
   preset: demo1
-  date_created: '20200416_191106'
+  date_created: '20200417_122433'
 preprocessing:
 - create_mask:
     tool: polygon
