+++ tests/resources/project1/data/0__stickleback_side/attributes.yaml
@@ -1,12 +1,0 @@
-image:
-  filename: stickleback_side.jpg
-  filepath: tutorials/images\stickleback_side.jpg
-  filetype: .jpg
-  width: 3000
-  height: 2000
-  size_ratio_original: 1
-project:
-  dirname: 0__stickleback_side
-  dirpath: tests/resources/project1\data\0__stickleback_side
-  raw_mode: link
-  raw_path: tutorials/images\stickleback_side.jpg
