+++ tests/resources/project2/attributes.yaml
@@ -1,0 +1,4 @@
+date_created: '20200417_151506'
+date_changed: '20200417_151506'
+root_dir: tests/resources/project2
+data_dir: tests/resources/project2/data
