+++ tests/resources/project1/attributes.yaml
@@ -1,4 +1,4 @@
-date_created: '20200417_153727'
-date_changed: '20200417_153727'
+date_created: '20200417_155015'
+date_changed: '20200417_155015'
 root_dir: tests/resources/project1
 data_dir: tests/resources/project1/data
