+++ tests/resources/project1/data/0__stickle2/pype_config_v1.yaml
@@ -8,7 +8,7 @@ image:
 pype:
   name: v1
   preset: demo1
-  date_created: '20200417_155015'
+  date_created: '20200417_163301'
 preprocessing:
 - create_mask:
     tool: polygon
