+++ tests/resources/project1/attributes.yaml
@@ -1,4 +1,4 @@
-date_created: '20200417_163301'
-date_changed: '20200417_163301'
+date_created: '20200417_164933'
+date_changed: '20200417_164933'
 root_dir: tests/resources/project1
 data_dir: tests/resources/project1/data
