+++ tests/resources/project1/attributes.yaml
@@ -1,4 +1,4 @@
-date_created: '20200417_173000'
-date_changed: '20200417_173000'
+date_created: '20200417_174145'
+date_changed: '20200417_174145'
 root_dir: tests/resources/project1
 data_dir: tests/resources/project1\data
