+++ tests/resources/project2/data/0__stickleback_side/pype_config_v1.yaml
@@ -1,4 +1,4 @@
-    image:
+image:
   filename: stickleback_side.jpg
   filepath: tutorials/images/stickleback_side.jpg
   filetype: .jpg
