+++ tests/resources/project2/data/0__stickle1/pype_config_v1.yaml
@@ -1,6 +1,6 @@
 image:
   filename: stickle1.JPG
-  filepath: tutorials/images/stickle1.JPG
+  filepath: tutorials/images\stickle1.JPG
   filetype: .JPG
   width: 2400
   height: 1600
@@ -8,7 +8,7 @@ image:
 pype:
   name: v1
   preset: demo1
-  date_created: '20200417_151506'
+  date_created: '20200417_190739'
 preprocessing:
 - create_mask:
     tool: polygon
