+++ tests/resources/project2/data/0__stickle1/attributes.yaml
@@ -1,15 +1,15 @@
 image:
   filename: stickle1.JPG
-  filepath: tutorials/images/stickle1.JPG
+  filepath: tutorials/images\stickle1.JPG
   filetype: .JPG
   width: 2400
   height: 1600
   size_ratio_original: 1
 project:
   dirname: 0__stickle1
-  dirpath: tests/resources/project2/data/0__stickle1
+  dirpath: tests/resources/project2\data\0__stickle1
   raw_mode: link
-  raw_path: tutorials/images/stickle1.JPG
+  raw_path: tutorials/images\stickle1.JPG
 scale:
-  template_path: tests/resources/project2/scale_template.jpg
-  template_px_mm_ratio: 35
+  template_path: tests/resources/project2\scale_template.jpg
+  template_px_mm_ratio: 36
