### 2020-04-18 #23 @mluerig
@lwasser Phenopype is almost in ok shape to be reviewed. However, one module (video-analysis, which is an extension of the core image analysis kit) is still not working great and I would like to spend a few more weeks with it. It's an important, but code-wise, peripheral feature. Would you agree to a review of the program as it is now (linux builds passing, 70% coverage, all docstrings are there, tutorials and vignettes as well), and then I add the fully functional video-analysis module to a later re-submission?
