+++ tests/resources/project1/data/0__stickle2/pype_config_v1.yaml
@@ -1,45 +1,0 @@
-image:
-  filename: stickle2.JPG
-  filepath: tutorials/images\stickle2.JPG
-  filetype: .JPG
-  width: 2400
-  height: 1600
-  size_ratio_original: 1
-pype:
-  name: v1
-  preset: demo1
-  date_created: '20200417_190738'
-preprocessing:
-- create_mask:
-    tool: polygon
-segmentation:
-- threshold:
-    method: adaptive
-    blocksize: 199
-    constant: 5
-    channel: red
-- morphology:
-    operation: close
-    shape: ellipse
-    kernel_size: 3
-    iterations: 3
-- find_contours:
-    retrieval: ext
-    min_diameter: 0
-    min_area: 150
-measurement:
-visualization:
-- select_canvas:
-    canvas: image
-- draw_contours:
-    line_width: 2
-    label_width: 1
-    label_size: 1
-    fill: 0.3
-- draw_masks
-export:
-- save_contours:
-    overwrite: true
-- save_canvas:
-    resize: 0.5
-    overwrite: true
