+++ tests/resources/project1/attributes.yaml
@@ -1,4 +1,0 @@
-date_created: '20200417_190738'
-date_changed: '20200417_190738'
-root_dir: tests/resources/project1
-data_dir: tests/resources/project1\data
