### 2020-04-20 #23 @mluerig
ok I'll finish it up before then.

one more thing about CI: I recently started using travis CI with my program, and I discovered i) that travis doesn't support python builds for windows and macOS, and ii) testing some of the functions is tricky because they open up a GUI requiring user interaction.

I can't change i), and I tried to mimic as much user input for ii) as possible, but ultimately, I will have to run the tests locally to get fully coverage. is this acceptable?
