### 2020-04-20 #23 @lwasser
hi @mluerig these are both good questions
re travis. there are a few options. I believe in our [cookiecutter example](https://github.com/pyOpenSci/cookiecutter-pyopensci) we have an example implementation for appveyor  which runs windows. circleci also has windows options now (i haven't tried it yet!). But there are several windows options available. we've been using appveyor for our packages.

Re the gui implementation for tests... i need to dig a little bit more into that. let me see what folks say on the discourse forum / twitter and i will get back to you. Ideally tests can be all run via ci. it seems like you could potentially implement some sort of monkey patching to mimic  user input but i really don't know enough about this to make any suggestions! more to come on this.
