### 2020-04-20 #23 @mluerig
@lwasser ah I didn't know about appveyor - I'll check it out

yeah I have been monkeypatching keyboard-input using the mock package, that works great. but clicking into an image is something else. I think I could get GUI functions working on CI by supplying some default coordinates and timers, but this would require a special testing interface for some of the functions, which I would implement as a last resort
