### 2020-04-20 #23 @lwasser
@mluerig yes appveyor works pretty well!

i can totally get how clicking on an image would be difficult to recreate. let me do a bit of digging and i'll get back to you! specifically what is the user input providing? aoi regions of the image to analyze or training data or something to that effect?
