### 2020-04-20 #23 @mluerig
yupp sometimes a mask needs to be selected to detect blobs within or a reference card measured to for automatic detection. it's not unimportant, but again, if we don't find anything I'll build an appropriate testing interface
