### 2020-04-20 ddaf6a0
tut 4 / df fix find_scale
