+++ docs/_sources/tutorial_3_managing_projects_1.ipynb.txt
@@ -45,21 +45,60 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 3,
    "metadata": {},
    "outputs": [],
    "source": [
     "import phenopype as pp\n",
-    "import os"
+    "import os\n",
+    "os.getcwd()"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 6,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "--------------------------------------------\n",
+      "Phenopype will create a new project at\n",
+      "\n",
+      "E:\\git_repos\\phenopype\\_temp\\my_project\\phenopype\n",
+      "\n",
+      "and change the current working directory to this location.\n",
+      "Proceed? (y/n)\n",
+      "y\n",
+      "\n",
+      "project attributes written to E:\\git_repos\\phenopype\\_temp\\my_project\\phenopype\\attributes.yaml\n",
+      "--------------------------------------------\n"
+     ]
+    }
+   ],
    "source": [
-    "myproj = pp.project(root_dir=r\"../_temp/my_project/phenopype_analysis\") ## doesn't have to be \"myproj\", can be named anything"
+    "myproj = pp.project(root_dir=r\"../_temp/my_project/phenopype\") ## doesn't have to be \"myproj\", can be named anything"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 9,
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "'E:\\\\git_repos\\\\phenopype\\\\_temp\\\\my_project\\\\phenopype'"
+      ]
+     },
+     "execution_count": 9,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
+   "source": [
+    "os.getcwd()"
    ]
   },
   {
@@ -71,11 +110,32 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 16,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "data": {
+      "text/plain": [
+       "['isopods.jpg',\n",
+       " 'isopods_fish.mp4',\n",
+       " 'phytoplankton.jpg',\n",
+       " 'snail.jpg',\n",
+       " 'stickle1.JPG',\n",
+       " 'stickle2.JPG',\n",
+       " 'stickle3.JPG',\n",
+       " 'stickleback_side.jpg',\n",
+       " 'stickleback_top.jpg',\n",
+       " 'worms.jpg']"
+      ]
+     },
+     "execution_count": 16,
+     "metadata": {},
+     "output_type": "execute_result"
+    }
+   ],
    "source": [
-    "os.listdir(\"images\") ## called from phenopype-master"
+    "images = \"../../../tutorials/images\"\n",
+    "os.listdir(images) ##"
    ]
   },
   {
@@ -87,11 +147,36 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 17,
    "metadata": {},
-   "outputs": [],
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "--------------------------------------------\n",
+      "phenopype will search for files at\n",
+      "\n",
+      "E:\\git_repos\\phenopype\\tutorials\\images\n",
+      "\n",
+      "using the following settings:\n",
+      "\n",
[…]
