+++ docs/_sources/index.rst.txt
@@ -68,8 +68,8 @@ Examples
 			<div class="thumbnail">
 			<a href="example_1_detect_objects_isopods.html">
 			<img src="_images/thumbs/ex1.jpg" alt="example_1_detect_objects_isopods" style="width:100%">
-			<b><p style="text-decoration: underline;">Detecting invertebrates</p></b>
-			<b>Organism:</b> freshwater isopods <br>
+			<b><p style="text-decoration: underline;">Detecting aquatic isopods</p></b>
+			<b>Organism:</b> waterlouse or sowbug<br>
 			<b>Species:</b> <i>Asellus aquaticus</i> <br>
 			<b>Traits:</b> size, pigmentation, body-shape <br>
 			</a>
@@ -135,6 +135,7 @@ Examples
 			<b>Organism:</b> freshwater snails <br>
 			<b>Species:</b> <i></i> <br>
 			<b>Traits:</b> size (area), abundance<br>
+			<br>
 			</a>
 			</div>
 		</div>
@@ -147,10 +148,10 @@ Examples
 			<div class="thumbnail">
 			<a href="example_7_worm_length.html">
 			<img src="_images/thumbs/ex7.jpg" alt="example_7_worm_length" style="width:100%">
-			<b><p style="text-decoration: underline;">Measure worm length</p></b>
-			<b>Organism:</b>  <br>
-			<b>Species:</b> <i></i> <br>
-			<b>Traits:</b> length, area<br>
+			<b><p style="text-decoration: underline;">Building a worm classifier using length-width ratio</p></b>
+			<b>Organism:</b>California blackworm<br>
+			<b>Species:</b> <i>Lumbriculus variegatus</i> <br>
+			<b>Traits:</b> length, width, area<br>
 			</a>
 			</div>
 		</div>
