+++ docs/_sources/index.rst.txt
@@ -132,10 +132,9 @@ Examples
 			<a href="example_6_counting_snails.html">
 			<img src="_images/thumbs/ex6.jpg" alt="example_6_counting_snails" style="width:100%">
 			<b><p style="text-decoration: underline;">Counting snails</p></b>
-			<b>Organism:</b> freshwater snails <br>
-			<b>Species:</b> <i></i> <br>
+			<b>Organism:</b> New Zealand mud snail <br>
+			<b>Species:</b> <i>Potamopyrgus antipodarum</i> <br>
 			<b>Traits:</b> size (area), abundance<br>
-			<br>
 			</a>
 			</div>
 		</div>
@@ -148,7 +147,7 @@ Examples
 			<div class="thumbnail">
 			<a href="example_7_worm_length.html">
 			<img src="_images/thumbs/ex7.jpg" alt="example_7_worm_length" style="width:100%">
-			<b><p style="text-decoration: underline;">Building a worm classifier using length-width ratio</p></b>
+			<b><p style="text-decoration: underline;">Measuring length-width ratios in worms</p></b>
 			<b>Organism:</b>California blackworm<br>
 			<b>Species:</b> <i>Lumbriculus variegatus</i> <br>
 			<b>Traits:</b> length, width, area<br>
