+++ docs/_sources/example_2_landmarks_stickleback.ipynb.txt
@@ -68,7 +68,7 @@
    "source": [
     "## just place a few test-landmarks\n",
     "\n",
-    "pp.measurement.landmarks(ct, point_size=5, point_col=\"green\", label_size=1, overwrite=True) \n",
+    "pp.measurement.landmarks(ct, point_size=15, point_colour=\"green\", label_size=1, overwrite=True) \n",
     "\n",
     "## sets landmarks. the landmarks get stored inside the container. if you have already placed landmarks\n",
     "## to the same container and set \"overwrite=False\", you wont be able to place them again until you set\n",
@@ -79,7 +79,7 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "To create an control-image with the selected points, the `draw_landmarks` has to be called explicitly. This is necessary when using the low throughput, but not the high throughput routine (see [Tutorial 2](tutorial_2_phenopype_workflow.ipynb)); likewise the `save_landmarks` function."
+    "To create an image with the selected points, we first need te select the background image (or \"canvas\") with `select_canvas` and then call the `draw_landmarks` function. These explicit steps are necessary when using the low throughput, but not the high throughput routine (see [Tutorial 2](tutorial_2_phenopype_workflow.ipynb)); likewise the `save_landmarks` function."
    ]
   },
   {
@@ -87,6 +87,13 @@
    "execution_count": 3,
    "metadata": {},
    "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "- raw image\n"
+     ]
+    },
     {
      "data": {
       "text/html": [
@@ -125,8 +132,8 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>1</td>\n",
-       "      <td>1034</td>\n",
-       "      <td>451</td>\n",
+       "      <td>739</td>\n",
+       "      <td>355</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>1</th>\n",
@@ -135,8 +142,8 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>2</td>\n",
-       "      <td>1202</td>\n",
-       "      <td>326</td>\n",
+       "      <td>1320</td>\n",
+       "      <td>259</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>2</th>\n",
@@ -145,8 +152,8 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>3</td>\n",
-       "      <td>1348</td>\n",
-       "      <td>473</td>\n",
+       "      <td>1368</td>\n",
+       "      <td>353</td>\n",
        "    </tr>\n",
        "  </tbody>\n",
        "</table>\n",
@@ -154,9 +161,9 @@
       ],
       "text/plain": [
        "       filename  width  height  size_ratio_original  landmark     x    y\n",
-       "0  stickle1.jpg   2400    1600                    1         1  1034  451\n",
-       "1  stickle1.jpg   2400    1600                    1         2  1202  326\n",
-       "2  stickle1.jpg   2400    1600                    1         3  1348  473"
+       "0  stickle1.jpg   2400    1600                    1         1   739  355\n",
+       "1  stickle1.jpg   2400    1600                    1         2  1320  259\n",
+       "2  stickle1.jpg   2400    1600                    1         3  1368  353"
       ]
      },
      "execution_count": 3,
@@ -165,7 +172,8 @@
     }
    ],
    "source": [
[…]
