### 2020-04-27 6a44f04
tut 6 /consec mask fix
