+++ docs/_sources/example_2_landmarks_stickleback.ipynb.txt
@@ -54,21 +54,21 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 2,
+   "execution_count": 6,
    "metadata": {},
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "- setting landmarks\n"
+      "- setting landmarks (overwriting)\n"
      ]
     }
    ],
    "source": [
     "## just place a few test-landmarks\n",
     "\n",
-    "pp.measurement.landmarks(ct, point_size=5, point_col=\"green\", label_size=1, overwrite=True) \n",
+    "pp.measurement.landmarks(ct, point_size=5, point_colour=\"green\", label_size=1, overwrite=True) \n",
     "\n",
     "## sets landmarks. the landmarks get stored inside the container. if you have already placed landmarks\n",
     "## to the same container and set \"overwrite=False\", you wont be able to place them again until you set\n",
@@ -84,9 +84,16 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 8,
    "metadata": {},
    "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "- raw image\n"
+     ]
+    },
     {
      "data": {
       "text/html": [
@@ -125,8 +132,8 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>1</td>\n",
-       "      <td>1034</td>\n",
-       "      <td>451</td>\n",
+       "      <td>792</td>\n",
+       "      <td>273</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>1</th>\n",
@@ -135,8 +142,8 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>2</td>\n",
-       "      <td>1202</td>\n",
-       "      <td>326</td>\n",
+       "      <td>1142</td>\n",
+       "      <td>338</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>2</th>\n",
@@ -145,8 +152,18 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>3</td>\n",
-       "      <td>1348</td>\n",
-       "      <td>473</td>\n",
+       "      <td>1212</td>\n",
+       "      <td>497</td>\n",
+       "    </tr>\n",
+       "    <tr>\n",
+       "      <th>3</th>\n",
+       "      <td>stickle1.jpg</td>\n",
+       "      <td>2400</td>\n",
+       "      <td>1600</td>\n",
+       "      <td>1</td>\n",
+       "      <td>4</td>\n",
+       "      <td>1344</td>\n",
+       "      <td>723</td>\n",
        "    </tr>\n",
        "  </tbody>\n",
        "</table>\n",
@@ -154,19 +171,21 @@
       ],
       "text/plain": [
        "       filename  width  height  size_ratio_original  landmark     x    y\n",
-       "0  stickle1.jpg   2400    1600                    1         1  1034  451\n",
-       "1  stickle1.jpg   2400    1600                    1         2  1202  326\n",
-       "2  stickle1.jpg   2400    1600                    1         3  1348  473"
+       "0  stickle1.jpg   2400    1600                    1         1   792  273\n",
+       "1  stickle1.jpg   2400    1600                    1         2  1142  338\n",
+       "2  stickle1.jpg   2400    1600                    1         3  1212  497\n",
+       "3  stickle1.jpg   2400    1600                    1         4  1344  723"
       ]
      },
-     "execution_count": 3,
+     "execution_count": 8,
[…]
