+++ docs/_sources/tutorial_6_video_analysis.ipynb.txt
@@ -6,7 +6,2588 @@
    "source": [
     "# Tutorial 6: Video analysis\n",
     "\n",
-    "[ADD TEXT]\n"
+    "The principles of video analysis and image analysis are similar. The phenopype video tracking module performs image segmentation using single video frames. However, when working with animals, a big advantage to static images is the fact that they move so that the differences among imges can be used to train a background model from which the foreground can segmented. This is called foreground-background subtraction (https://docs.opencv.org/3.4/d1/dc5/tutorial_background_subtraction.html).\n",
+    "\n",
+    "Currently the tracking module works standalone and uses only few routines of the phenopype core modules. Future releases may feature a better integration with the different phenopype workflows.  "
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "<center>\n",
+    "<div class=\"row; text-align: left\">\n",
+    "<div class=\"col-md-6\">\n",
+    "    \n",
+    "![](_assets/tracking1.gif)\n",
+    "    \n",
+    "</div>\n",
+    "<div class=\"col-md-6\">\n",
+    "\n",
+    "![](_assets/tracking2.png)\n",
+    "\n",
+    "</div>\n",
+    "    \n",
+    "**Fig. 1:** The motion tracker module enables the tracking of multiple objects that can be distinguished by size. For example, it is possible to record and separate the movement of both fish (red) and all 20 isopods (green) in this predator-prey interaction experiment. \n",
+    "    \n",
+    "</div>\n",
+    "</center>"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## The motion tracker class\n",
+    "\n",
+    "First, the `motion_tracker` class needs to be called on the video-path. This will initialize the procedure with basic information on the input video. Specifying video output is optional: the tracking procedure can be run without saving any record of the tracking-results, but only the results (i.e. just a csv with coordinates and no video). Should the `video_ouput` method be called then it will save an output video with similar properties (codec, dimensions, fps.) of the input file.    "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import phenopype as pp\n",
+    "import os"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "video_path =  r\"images/isopods_fish.mp4\"\n",
+    "out_dir =  r\"../_temp/video\""
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "\n",
+      "\n",
[…]
