+++ docs/_sources/example_2_landmarks_stickleback.ipynb.txt
@@ -6,7 +6,7 @@
    "source": [
     "# Example 2: Stickleback morphometrics - landmarks\n",
     "\n",
-    "Functional morphology of organisms is often measured by placing landmarks at specific points that show structural, functional or developmental significance. In this example phenopype is used to place morphometric landmarks across the anterior half of a stickleback (*Gasterosteus aculeatus*) stained with alizarin red.\n",
+    "Functional morphology of organisms is often measured by placing landmarks at specific points that show structural, functional or developmental significance. In this example phenopype is used to place morphometric landmarks across the anterior half of a threespine stickleback (*Gasterosteus aculeatus*) stained with alizarin red.\n",
     "\n",
     "First we place landmarks in low throughput mode to learn how the landmark-function works, then we look at a high throughput landkmark example with a project directory and a global scale. "
    ]
