+++ docs/_sources/api.rst.txt
@@ -167,20 +167,22 @@ Export
 Video analysis
 --------------
 
-Generic motion tracking
-"""""""""""""""""""""""
+.. raw:: html
 
-[ADD TEXT]
+	These classes provide basic motion tracking features via foreground-background subtraction - for more details see 
+	<a href="https://docs.opencv.org/3.4/d1/dc5/tutorial_background_subtraction.html">the OpenCV docs 2</a> and 
+	<a href="tutorial_6_video_analysis.html">Tutorial 6</a>.
+
+Motion tracker
+""""""""""""""
 
 .. autoclass:: phenopype.tracking.motion_tracker
 	:members:
 	:undoc-members:
 	:show-inheritance:	
 
-Define tracking method
-""""""""""""""""""""""
-
-[ADD TEXT]
+Tracking methods
+""""""""""""""""
 
 .. autoclass:: phenopype.tracking.tracking_method
 	:members:
