+++ docs/_sources/example_7_worm_length.ipynb.txt
@@ -60,7 +60,7 @@
     "import phenopype as pp\n",
     "\n",
     "image_path = r\"./images/worms.jpg\"\n",
-    "custom_save_path = r\"../_temp/output\"\n",
+    "custom_save_path = r\"../_temp/output/ex7\"\n",
     "ct = pp.load_image(image_path, # path to image\n",
     "                   cont=True,  # create a phenopype container\n",
     "                   save_suffix=\"worms1\",      # indicate a suffix for save- and loading operations\n",
@@ -76,10 +76,7 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "AUTOLOAD\n",
-      "- drawing loaded from attributes.yaml\n",
-      "- polylines_worms1.csv\n",
-      "- masks_worms1.csv\n"
+      "Nothing loaded.\n"
      ]
     }
    ],
@@ -89,7 +86,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 4,
    "metadata": {},
    "outputs": [],
    "source": [
@@ -105,14 +102,14 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 4,
+   "execution_count": 5,
    "metadata": {},
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "- mask with label mask1 already created (overwrite=False)\n"
+      "- create mask\n"
      ]
     }
    ],
@@ -131,7 +128,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 6,
    "metadata": {},
    "outputs": [
     {
@@ -164,7 +161,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 7,
    "metadata": {},
    "outputs": [
     {
@@ -189,16 +186,16 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 7,
+   "execution_count": 8,
    "metadata": {},
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "- masks saved under ../_temp/output\\masks_worms1.csv (overwritten).\n",
-      "- contours saved under ../_temp/output\\contours_worms1.csv (overwritten).\n",
-      "- canvas saved under ../_temp/output\\canvas_worms1.jpg (overwritten).\n"
+      "- masks saved under ../_temp/output\\masks_worms1.csv.\n",
+      "- contours saved under ../_temp/output\\contours_worms1.csv.\n",
+      "- canvas saved under ../_temp/output\\canvas_worms1.jpg.\n"
      ]
     }
    ],
@@ -221,12 +218,20 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "Should a worm superimpose itself it is not possible for the skeletonization algorithm to follow the shape through from beginning to end (e.g. in worm 3 in this example). In that case one has to manually draw a line along the worm using the `polylines` tool. "
+    "Should a worm superimpose itself it is not possible for the skeletonization algorithm to follow the shape through from beginning to end (e.g. in worm 3 in this example). In that case one has to manually draw a line along the worm using the `polylines` tool. \n",
+    "\n",
+    "<center>\n",
+    "<div style=\"width:300px; text-align: left\" >\n",
+    "    \n",
+    "![Canvas masking](_assets/ex7_poly.png)\n",
+    "    \n",
+    "</div>\n",
+    "</center>"
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 10,
+   "execution_count": 11,
[…]
