+++ docs/_sources/example_2_landmarks_stickleback.ipynb.txt
@@ -6,7 +6,7 @@
    "source": [
     "# Example 2: Stickleback morphometrics - landmarks\n",
     "\n",
-    "Functional morphology of organisms is often measured by placing landmarks at specific points that show structural, functional or developmental significance. In this example phenopype is used to place morphometric landmarks across the anterior half of a threespine stickleback (*Gasterosteus aculeatus*) stained with alizarin red.\n",
+    "Functional morphology of organisms is often measured by placing landmarks at specific points that show structural, functional or developmental significance. In this example phenopype is used to place morphometric landmarks across the anterior half of a stickleback (*Gasterosteus aculeatus*) stained with alizarin red.\n",
     "\n",
     "First we place landmarks in low throughput mode to learn how the landmark-function works, then we look at a high throughput landkmark example with a project directory and a global scale. "
    ]
@@ -54,21 +54,24 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 6,
+   "execution_count": 2,
    "metadata": {},
    "outputs": [
     {
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "- setting landmarks (overwriting)\n"
+      "- setting landmarks\n"
      ]
     }
    ],
    "source": [
     "## just place a few test-landmarks\n",
     "\n",
-    "pp.measurement.landmarks(ct, point_size=5, point_colour=\"green\", label_size=1, overwrite=True) \n",
+    "pp.measurement.landmarks(ct, \n",
+    "                         point_size=15, \n",
+    "                         point_colour=\"green\", \n",
+    "                         label_size=1, overwrite=True) \n",
     "\n",
     "## sets landmarks. the landmarks get stored inside the container. if you have already placed landmarks\n",
     "## to the same container and set \"overwrite=False\", you wont be able to place them again until you set\n",
@@ -79,12 +82,12 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "To create an control-image with the selected points, the `draw_landmarks` has to be called explicitly. This is necessary when using the low throughput, but not the high throughput routine (see [Tutorial 2](tutorial_2_phenopype_workflow.ipynb)); likewise the `save_landmarks` function."
+    "To create an image with the selected points, we first need te select the background image (or \"canvas\") with `select_canvas` and then call the `draw_landmarks` function. These explicit steps are necessary when using the low throughput, but not the high throughput routine (see [Tutorial 2](tutorial_2_phenopype_workflow.ipynb)); likewise the `save_landmarks` function."
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 8,
+   "execution_count": 3,
    "metadata": {},
    "outputs": [
     {
@@ -132,8 +135,8 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>1</td>\n",
-       "      <td>792</td>\n",
-       "      <td>273</td>\n",
+       "      <td>1240</td>\n",
+       "      <td>379</td>\n",
        "    </tr>\n",
        "    <tr>\n",
        "      <th>1</th>\n",
@@ -142,8 +145,8 @@
        "      <td>1600</td>\n",
        "      <td>1</td>\n",
        "      <td>2</td>\n",
-       "      <td>1142</td>\n",
[…]

[…]
