+++ docs/_sources/tutorial_5_gui_interactions.ipynb.txt
@@ -6,7 +6,53 @@
    "source": [
     "# Tutorial 5: GUI interactions (masks, lines and drawing tools)\n",
     "\n",
-    "Phenopype "
+    "Phenopype uses OpenCV's [High GUI module](https://docs.opencv.org/3.4/d7/dfc/group__highgui.html) to display images and to allow users to interact with images. High GUI has a few pros and cons: \n",
+    "\n",
+    "```\n",
+    "+ native OpenCV GUI (no extra GUI libraries required)\n",
+    "+ the module is extremely fast in displaying images \n",
+    "+ it can display very large image-arrays (> 10000x10000 pixels)\n",
+    "+ it can display multiple images side by side\n",
+    "+ interactions (drawing and measuring) are possible \n",
+    "\n",
+    "- sometimes unstable (e.g. windows are not closed but freeze)\n",
+    "- issues with cross plattform stability (e.g. on macOS)\n",
+    "- displaying instructions is hacky (text is \"painted\" onto a displayed image) \n",
+    "- user input (key strokes and mouse clicks) sometimes isn't captured properly\n",
+    "```\n",
+    "\n",
+    "Currently Phenopype uses the standard GUI libraries that ship with the most recent precompiled `opencv-contrib-python` that are listed on the [Python package index](https://pypi.org/project/opencv-python/), which is `Qt` for Linux and macOS, and `Win32 UI` on Windows. The `Qt` GUI is a bit more userfriendly with builtin buttons, scrollbars, RGB info and zoom ([see the OpenCV docs](https://docs.opencv.org/master/dc/d46/group__highgui__qt.html)), but you don't actually need those things for Phenopype GUI interactions. "
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "**Notes regarding High GUI windows in Phenopype** \n",
+    "\n",
+    "- If you want to close a High GUI window, use `Enter`\n",
+    "- If you want to close a High GUI window and quit the Phenoype process that invoked it, use `Esc`\n",
+    "- If your High GUI window is frozen, use `Esc`. If the window still does't close, type `cv2.destroyAllWindows()` into the console\n",
+    "- If the last step doesn't work, restart the kernel"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "## Show images\n",
+    "\n",
+    "To show an image, it first has to be loaded as an array using `load_image`. The array can then be passed on to `show_image`, which simply displays an image (no interactions, except zooming in using the mousewheel). The window is closed by keystroke (`Enter`, or `Esc` which closes the window and ends ongoing processes). "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import phenopype as pp\n",
+    "import os"
    ]
   },
   {
@@ -15,14 +61,434 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "## Show images\n",
-    "\n",
+    "img = pp.load_image(\"images/stickle1.jpg\") \n",
+    "pp.show_image(img,\n",
+    "              window_aspect=\"free\" # resize the window\n",
+    "             )"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "However, `show_image` can also handle multiple arrays. Here we loop through the *images* folder, attach all images to a list, and pass that list of arrays to the functions - it will give a warning if more than 10 images are being opened at the same time. "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
[…]
