+++ docs/_sources/example_6_counting_snails.ipynb.txt
@@ -41,7 +41,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 3,
+   "execution_count": 1,
    "metadata": {},
    "outputs": [
     {
@@ -73,7 +73,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": 4,
+   "execution_count": 2,
    "metadata": {},
    "outputs": [
     {
@@ -94,12 +94,12 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "Draw a mask around the snails inside the tray by dragging a rectangle around them - finish with `Enter`, abort with `Esc`."
+    "Begin by drawing a mask around the snails inside the tray by dragging a rectangle around them - finish with `Enter`, abort with `Esc`."
    ]
   },
   {
    "cell_type": "code",
-   "execution_count": 5,
+   "execution_count": 3,
    "metadata": {},
    "outputs": [
     {
@@ -118,7 +118,60 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "Now we'll blur the image a little, convert it to a binary image, and look at the results:"
+    "Now we need to measure the pixel-to-mm-ratio. The snail pictures don't contain a scale that is appropriate for automatic detection - the millimeter-paper on the side does not contain enough unique keypoints to be detected by the [AKAZE algorithm](http://www.bmva.org/bmvc/2013/Papers/paper0013/abstract0013.pdf) that powers `find_scale`. Therefore each image needs to be manually measured by dragging a line across the mm-paper, and entering the distance. \n",
+    "\n",
+    "**Note:** It is better to measure a long distance than a short distance to minimize the measurement error - use the full length (~70 mm) here."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 4,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "- measure pixel-to-mm-ratio\n",
+      "Scale set\n",
+      "- add column length\n",
+      "- scale pixel-to-mm-ratio already measured (overwrite=False)\n"
+     ]
+    }
+   ],
+   "source": [
+    "pp.preprocessing.create_scale(ct)"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "For high data quality it is important to verify the ID of the specimen in the current picture. Often, the picture name contains the ID, but typically an label is placed inside the image. Using the `enter_data` tool we open the image and an entry prompt that will create a column with a name of our chosing inside all exported results."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 5,
+   "metadata": {},
+   "outputs": [
+    {
+     "name": "stdout",
+     "output_type": "stream",
+     "text": [
+      "- add column ID\n"
+     ]
+    }
+   ],
+   "source": [
+    "pp.preprocessing.enter_data(ct, columns=\"ID\")"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Next is segmentation. First we blur the image a little, convert it to a binary image, and look at the results:"
    ]
   },
   {
@@ -139,7 +192,7 @@
     "pp.segmentation.threshold(ct, \n",
     "                          method=\"adaptive\", \n",
     "                          blocksize=59, ## relatively low sensitivity\n",
-    "                          constant=10) ## relatively high subtraction factor from the result\n",
[…]
