### 2020-05-04 #23 @mluerig
closing - actual submission here: https://github.com/pyOpenSci/software-review/issues/24

@lwasser let me know if you would like to receive my suggestions for reviewers
